backend/dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=dev --output-file=dev-requirements.txt pyproject.toml
#
asgiref==3.7.2
    # via django
astor==0.8.1
    # via flake8-simplify
astroid==2.15.7
    # via flake8-django
attrs==23.1.0
    # via
    #   flake8-bugbear
    #   flake8-eradicate
    #   flake8-multiline-containers
black==23.9.1
    # via flake8-black
click==8.1.7
    # via black
cognitive-complexity==1.3.0
    # via flake8-cognitive-complexity
colorama==0.4.6
    # via click
django==4.2.5
    # via
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   drum-score-backend (pyproject.toml)
django-cors-headers==4.2.0
    # via drum-score-backend (pyproject.toml)
django-environ==0.11.2
    # via drum-score-backend (pyproject.toml)
django-filter==23.3
    # via drum-score-backend (pyproject.toml)
djangorestframework==3.14.0
    # via drum-score-backend (pyproject.toml)
eradicate==2.3.0
    # via flake8-eradicate
flake8==6.1.0
    # via
    #   flake8-absolute-import
    #   flake8-black
    #   flake8-bugbear
    #   flake8-django
    #   flake8-eradicate
    #   flake8-isort
    #   flake8-multiline-containers
    #   flake8-mutable
    #   flake8-pep3101
    #   flake8-print
    #   flake8-printf-formatting
    #   flake8-pyproject
    #   flake8-pytest
    #   flake8-simplify
    #   flake8-tidy-imports
    #   flake8-use-fstring
    #   flake8-walrus
flake8-absolute-import==1.0.0.1
    # via drum-score-backend (pyproject.toml)
flake8-black==0.3.6
    # via drum-score-backend (pyproject.toml)
flake8-bugbear==23.9.16
    # via drum-score-backend (pyproject.toml)
flake8-cognitive-complexity==0.1.0
    # via drum-score-backend (pyproject.toml)
flake8-django==1.4
    # via drum-score-backend (pyproject.toml)
flake8-eradicate==1.5.0
    # via drum-score-backend (pyproject.toml)
flake8-fixme==1.1.1
    # via drum-score-backend (pyproject.toml)
flake8-isort==6.1.0
    # via drum-score-backend (pyproject.toml)
flake8-multiline-containers==0.0.19
    # via drum-score-backend (pyproject.toml)
flake8-mutable==1.2.0
    # via drum-score-backend (pyproject.toml)
flake8-pep3101==2.0.0
    # via drum-score-backend (pyproject.toml)
flake8-pie==0.16.0
    # via drum-score-backend (pyproject.toml)
flake8-plugin-utils==1.3.3
    # via flake8-pytest-style
flake8-print==5.0.0
    # via drum-score-backend (pyproject.toml)
flake8-printf-formatting==1.1.2
    # via drum-score-backend (pyproject.toml)
flake8-pyproject==1.2.3
    # via drum-score-backend (pyproject.toml)
flake8-pytest==1.4
    # via drum-score-backend (pyproject.toml)
flake8-pytest-style==1.7.2
    # via drum-score-backend (pyproject.toml)
flake8-simplify==0.21.0
    # via drum-score-backend (pyproject.toml)
flake8-tidy-imports==4.10.0
    # via drum-score-backend (pyproject.toml)
flake8-todo==0.7
    # via drum-score-backend (pyproject.toml)
flake8-use-fstring==1.4
    # via drum-score-backend (pyproject.toml)
flake8-variables-names==0.0.6
    # via drum-score-backend (pyproject.toml)
flake8-walrus==1.2.0
    # via drum-score-backend (pyproject.toml)
gunicorn==21.2.0
    # via drum-score-backend (pyproject.toml)
isort==5.12.0
    # via flake8-isort
lazy-object-proxy==1.9.0
    # via astroid
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via
    #   black
    #   drum-score-backend (pyproject.toml)
    #   gunicorn
pathspec==0.11.2
    # via black
platformdirs==3.10.0
    # via black
pycodestyle==2.11.0
    # via
    #   flake8
    #   flake8-print
    #   flake8-todo
pyflakes==3.1.0
    # via flake8
pytz==2023.3.post1
    # via
    #   djangorestframework
    #   drum-score-backend (pyproject.toml)
sqlparse==0.4.4
    # via
    #   django
    #   drum-score-backend (pyproject.toml)
typing-extensions==4.8.0
    # via flake8-pie
tzdata==2023.3
    # via
    #   django
    #   drum-score-backend (pyproject.toml)
wrapt==1.15.0
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
